+++ arrays/src/main/kotlin/arrays/FuncoesBigDecimal.kt
@@ -1,0 +1,32 @@
+package arrays
+
+import java.math.BigDecimal
+import java.math.RoundingMode
+
+fun calculaAumentoRelativo(salario: BigDecimal, aumento: BigDecimal) =
+    if (salario < "5000.00".toBigDecimal()) {
+        salario + "500.00".toBigDecimal()
+    } else {
+        (salario * aumento).setScale(2, RoundingMode.UP)
+    }
+
+fun bigDecimalArrayOf(vararg valores: String): Array<BigDecimal> {
+    return Array(valores.size) { i ->
+        valores[i].toBigDecimal()
+    }
+}
+
+//Criação de uma Extension Function
+fun Array<BigDecimal>.somatoria(): BigDecimal {
+    return this.reduce { acumulador, valor ->
+        acumulador + valor
+    }
+}
+
+fun Array<BigDecimal>.media(): BigDecimal {
+    return if (this.isEmpty()) {
+        BigDecimal.ZERO
+    } else {
+        this.somatoria() / this.size.toBigDecimal()
+    }
+}
